refactor: move 6 backend modules into backends/
