+++ helix_context/backends/deberta_backend.py
@@ -29,7 +29,7 @@ from typing import Dict, List, Optional
 
 import torch
 
-from .schemas import Gene, NLRelation
+from ..schemas import Gene, NLRelation
 
 log = logging.getLogger("helix.ribosome.deberta")
