+++ scripts/backfill_bgem3_v2.py
@@ -36,7 +36,7 @@ import numpy as np
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
-from helix_context.bgem3_codec import BGEM3Codec
+from helix_context.backends.bgem3_codec import BGEM3Codec
 from helix_context.config import load_config
