+++ tests/test_deberta_backend.py
@@ -110,7 +110,7 @@ def test_deberta_rerank_chunks_in_recommended_batch_size(monkeypatch):
     calls (6 full + 1 partial of 4)."""
     _override_hardware(monkeypatch, rerank_batch=16)
 
-    from helix_context.deberta_backend import DeBERTaRibosome
+    from helix_context.backends.deberta_backend import DeBERTaRibosome
 
     rib = DeBERTaRibosome.__new__(DeBERTaRibosome)  # bypass __init__
     rib._device = torch.device("cpu")
@@ -134,7 +134,7 @@ def test_deberta_splice_chunks_in_recommended_batch_size(monkeypatch):
     expect ceil(50/16) = 4 tokenizer calls (16 + 16 + 16 + 2)."""
     _override_hardware(monkeypatch, splice_batch=16)
 
-    from helix_context.deberta_backend import DeBERTaRibosome
+    from helix_context.backends.deberta_backend import DeBERTaRibosome
 
     rib = DeBERTaRibosome.__new__(DeBERTaRibosome)  # bypass __init__
     rib._device = torch.device("cpu")
@@ -168,7 +168,7 @@ def test_deberta_init_consults_get_hardware_when_device_none(monkeypatch):
 
     _override_hardware(monkeypatch, rerank_batch=8)
 
-    import helix_context.deberta_backend as db
+    import helix_context.backends.deberta_backend as db
 
     monkeypatch.setattr(
         "transformers.AutoTokenizer.from_pretrained",
