+++ tests/test_density_gate.py
@@ -779,7 +779,7 @@ class TestColdTierRetrieval:
 
     @pytest.fixture(scope="class")
     def codec(self):
-        from helix_context.sema import SemaCodec
+        from helix_context.backends.sema import SemaCodec
         return SemaCodec()
 
     @pytest.fixture
